Use temporary variables to keep track of proceeded option
This is done because in case of error, optindex and nextchar pointer
could have changed, but we want to provide information about the option
that caused the error.

e.g. ./a.out -vi #v does not take any arg, but i does

In the above case, if we don't keep tmp variables, getopt will say that
v requires an argument (because before this commit we used optindex - 1
anyway)

Now we also check the saved nextchar to print error, so that it works
also with short options

Signed-off-by: Loïc Buckwell <lfalkau@student.42.fr>