Check opt struct members before dereferencing them
The for loop involved while searching for options
is running while at least one of the option style
(longname/shortname) in the opt_t element is not
NULL (or 0).

But if one was set, both were dereferenced to test
the next arg. If an invalid long option was given
in the program arguments, and an opt_t element has
only a short option, the long opt would have been
dereferenced, causing a segmentation fault.

This is now fixed.

Signed-off-by: lfalkau <lfalkau@student.42.fr>